// src/app/models/library.models.ts
export enum SelectionFilter {
  CLASS = 'CLASS',
  NEUTRAL = 'NEUTRAL',
  MINION = 'MINION',
  SPELL = 'SPELL',
  WEAPON = 'WEAPON',
  COST_0 = 'COST_0',
  COST_1 = 'COST_1',
  COST_2 = 'COST_2',
  COST_3 = 'COST_3',
  COST_4 = 'COST_4',
  COST_5 = 'COST_5',
  COST_6 = 'COST_6',
  COST_7_PLUS = 'COST_7_PLUS',
}